Force fade-in on finished uploads
Uploaded objects sometimes fail to fade in, especially on large folder
uploads. We tried to fix this previously in #804, but the class we need to fade
in is sometimes (maybe always?) called `template-upload` rather than
`template-download`, so the fix was incomplete. This may fix the various issues
with batch XML import.

See: #785, #804, #872, & #873.